refactor(infrastructure): enhance OAuth token validation and error handling
