+++ packages/infrastructure/src/contexts/user/services/index.ts
@@ -4,4 +4,4 @@ export * from './GoogleOAuthService';
 export * from './DiscordOAuthService';
 export * from './AesTokenEncryptionService';
 export * from './HtmlTemplateService';
-export * from './oauth.tokens';
+export * from './oauthTokens';
